+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Version 2.1.1
+- Added auto generated model definitions 
+- Prevent running the command in production env.
+- Fix null description and query param.
+- Allow user to customise consume param on APICall
+
 ## Version 2.1
 - Changed `ApiSuccessPaginated` to split the pagination data.
 - Documentation builder changed to create swagger and postman files separately.